feat(calidad): restaurar cobertura + corregir vulnerabilidades y bugs (nivel produccion)
- Externaliza credenciales de seed (resuelve 4 vulnerabilidades S6437)
- Corrige bug S2184 en ExcelExportService (cast a double)
- Restaura toda la suite de tests (cobertura backend ~98%, frontend ~63%)
- Mantiene fixes previos (leaked value JSX, Duration timezone-aware)

+++ backend/budget-service/src/main/java/com/ohchurus/budget/service/impl/ExcelExportService.java
@@ -213,7 +213,7 @@ public class ExcelExportService {
             String estado = Boolean.TRUE.equals(m.getConfirmed()) ? "Ejecutado" : "Presupuestado";
 
             XSSFCell cId = row.createCell(0);
-            cId.setCellValue(rowIdx - 1);
+            cId.setCellValue((double) (rowIdx - 1));
             cId.setCellStyle(centerStyle);
 
             XSSFCell cTipo = row.createCell(1);
